-split EventInformationItem to discreet file
